refactor(channels/discord): P1 — extract chunk.rs (message splitting)
Move split_message_for_discord{,_multi} and chunks_for_send into discord/chunk.rs
(deep leaf module; imports only DISCORD_MAX_MESSAGE_LENGTH from the contract
layer). Re-exported so existing callers and the test corpus resolve unchanged.
Behaviour-neutral: all 157 chunk/split/send tests pass.
